fix: id 통합

+++ app/src/main/java/com/hgh/na_o_man/data/dto/share_group/request/GroupJoinRequestDto.kt
@@ -1,6 +1,6 @@
 package com.hgh.na_o_man.data.dto.share_group.request
 
 data class GroupJoinRequestDto(
-    val profileId: Int,
+    val profileId: Long,
     val shareGroupId: Int
 )
