+++ app/src/main/java/com/hgh/na_o_man/data/dto/share_group/response/GroupJoinDto.kt
@@ -3,7 +3,7 @@ package com.hgh.na_o_man.data.dto.share_group.response
 import com.hgh.na_o_man.domain.model.share_group.GroupJoinModel
 
 data class GroupJoinDto(
-    val shareGroupId: Int
+    val shareGroupId: Long
 ) {
     fun toModel() = GroupJoinModel(
         shareGroupId = this.shareGroupId
