+++ app/src/main/java/com/hgh/na_o_man/domain/model/photo/PhotoUploadCountModel.kt
@@ -1,6 +1,6 @@
 package com.hgh.na_o_man.domain.model.photo
 
 data class PhotoUploadCountModel(
-    val shareGroupId: Int,
+    val shareGroupId: Long,
     val uploadCount: Int
 )
