+++ app/src/main/java/com/hgh/na_o_man/domain/model/share_group/GroupJoinModel.kt
@@ -1,5 +1,5 @@
 package com.hgh.na_o_man.domain.model.share_group
 
 data class GroupJoinModel(
-    val shareGroupId: Int
+    val shareGroupId: Long
 )
